server/requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sse-starlette>=1.8.0
beautifulsoup4>=4.12.0
httpx>=0.26.0
ddgs>=2.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
google-generativeai>=0.4.0
google-genai>=0.3.0
langchain-core>=0.1.0
langchain-core>=0.1.0
